requirements.txt
torch
torchmetrics
torchvision
torch-lightning
